+++ examples/snowflake-dagster-metadata-job/.env.example
@@ -1,0 +1,6 @@
+SNOWFLAKE_ACCOUNT=
+SNOWFLAKE_USERNAME=
+SNOWFLAKE_PASSWORD=
+SNOWFLAKE_DATABASE=
+SNOWFLAKE_WAREHOUSE=
+SNOWFLAKE_ROLE=
