+++ examples/snowflake-dagster-metadata-job/.gitignore
@@ -1,0 +1,4 @@
+.direnv/
+.envrc
+data/staging/*.txt
+tmp*/
