+++ examples/snowflake-dagster-metadata-job/pyproject.toml
@@ -1,0 +1,6 @@
+[build-system]
+requires = ["setuptools"]
+build-backend = "setuptools.build_meta"
+
+[tool.dagster]
+module_name = "snowflake_dagster_metadata_job"
