+++ examples/snowflake-dagster-metadata-job/setup.cfg
@@ -1,0 +1,2 @@
+[metadata]
+name = snowflake_dagster_metadata_job
