+++ examples/snowflake-dagster-metadata-job/snowflake_dagster_metadata_job/constants.py
@@ -1,0 +1,2 @@
+META_KEY_STORAGE_KIND = "dagster/storage_kind"
+META_KEY_STORAGE_IDENTIFIER = "dagster/relation_identifier"
