+++ examples/snowflake-dagster-metadata-job/snowflake_dagster_metadata_job/resources.py
@@ -1,0 +1,11 @@
+from dagster import EnvVar
+from dagster_snowflake import SnowflakeResource
+
+
+snowflake_resource = SnowflakeResource(
+    account=EnvVar("SNOWFLAKE_ACCOUNT"),
+    user=EnvVar("SNOWFLAKE_USER"),
+    password=EnvVar("SNOWFLAKE_PASSWORD"),
+    database=EnvVar("SNOWFLAKE_DATABASE"),
+    warehouse=EnvVar("SNOWFLAKE_WAREHOUSE"),
+)
